Conflicts:
	source/CMakeLists.txt
	source/Core/Base/src/platform/win32/Win32Window.cpp
	source/Core/Engine/include/Camera.h
	source/Core/Engine/include/Component.h
	source/Core/Engine/include/KeyCode.h
	source/Core/Engine/include/Moveable.h
	source/Core/Engine/local_include/Config.h.cmake
	source/Core/Engine/src/Component.cpp
	source/Core/Engine/src/Viewport.cpp
	source/Plugins/CMakeLists.txt

+++ source/Core/Engine/include/InputController.h
@@ -36,8 +36,8 @@ struct _NexEngineAPI InputControllerDesc {
 	static const InputControllerDesc Null;
 };
 
-/* ranges between 0 and max(AnalogValue)*/
-typedef int16 AnalogValue;
+/* ranges between 0 and 1*/
+typedef float AnalogValue;
 
 struct InputDir {
 	// x and y of delta
